+++ swdapp/src/lib/components/date-picker/datePicker.svelte
@@ -1,37 +1,18 @@
 <script lang="ts">
 	let weekDays: string[] = ["Mo", "Tu", "We", "Th", "Fr","Sa", "Su"];
 
-	const convertWeekDayName = {0: "Su", 1: "Mo", 2: "Tu", 3: "We", 4: "Th", 5: "Fr", 6: "Sa"};
-	const convertWeekToCalNum = {"Mo": 0, "Tu": 1, "We": 2, "Th": 3, "Fr": 4, "Sa": 5, "Su": 6};
-	const convertMonthNames = {"Jan": 0, "Feb": 1, "Mar": 2, "Apr": 3, "May": 4, "Jun": 5, "Jul": 6, "Aug": 7, "Sep": 8, "Oct": 9, "Nov": 10, "Dec": 11};
+	// const CAL_ROWS = 6;
+	// const CALC_COLS = 7;
+	// interface CalendarDay {placeholder: boolean, value?: string};
+	// let calDays:CalendarDay[][] =[];
 
-	function getMonthDayDetails(monthName: string, year: number) {
-		let month: number = (convertMonthNames as any)[monthName];
-		let date = new Date(year, month, 1);
+	// let calInfo = getMonthDayDetails("Jan", 2024);
+	// let startCalDayIdx = (convertWeekToCalNum as any)[calInfo.startDay];
+	// let sopCalDay = calInfo.numDays;
 
-		let beginDay = date.getDay();
-		let dayNum: number = 0;
-
-		while(date.getMonth() === month) {
-			dayNum = date.getDate();
-			date.setDate(date.getDate() + 1);
-		}
-
-		return {startDay: (convertWeekDayName as any)[beginDay], numDays: dayNum};
-	}
-
-	const CAL_ROWS = 6;
-	const CALC_COLS = 7;
-	interface CalendarDay {placeholder: boolean, value?: string};
-	let calDays:CalendarDay[][] =[];
-
-	let calInfo = getMonthDayDetails("Jan", 2024);
-	let startCalDayIdx = (convertWeekToCalNum as any)[calInfo.startDay];
-	let sopCalDay = calInfo.numDays;
-
-	for(let i = 0; i < CAL_ROWS; i++) {
-		let calRowDays:CalendarDay[] = Array(7).fill({placeholder: true});
-	}
+	// for(let i = 0; i < CAL_ROWS; i++) {
+	// 	let calRowDays:CalendarDay[] = Array(7).fill({placeholder: true});
+	// }
 </script>
 
 <!-- Datepicker -->
@@ -51,381 +32,9 @@
 			<!-- Weeks -->
 
 			<!-- Days -->
-			<div class="flex">
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-						disabled
-					>
-						26
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-						disabled
-					>
-						27
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-						disabled
-					>
-						28
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-						disabled
-					>
-						29
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-						disabled
-					>
-						30
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						1
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						2
-					</button>
-				</div>
-			</div>
+			<slot></slot>
 			<!-- Days -->
-
-			<!-- Days -->
-			<div class="flex">
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						3
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						4
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						5
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						6
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						7
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						8
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						9
-					</button>
-				</div>
-			</div>
-			<!-- Days -->
-
-			<!-- Days -->
-			<div class="flex">
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						10
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						11
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						12
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						13
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						14
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						15
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						16
-					</button>
-				</div>
-			</div>
-			<!-- Days -->
-
-			<!-- Days -->
-			<div class="flex">
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						17
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						18
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						19
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent bg-blue-600 text-sm font-medium text-white hover:border-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						20
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						21
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						22
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						23
-					</button>
-				</div>
-			</div>
-			<!-- Days -->
-
-			<!-- Days -->
-			<div class="flex">
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						24
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						25
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						26
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						27
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						28
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						29
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						30
-					</button>
-				</div>
-			</div>
-			<!-- Days -->
-
-			<!-- Days -->
-			<div class="flex">
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-					>
-						31
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-						disabled
-					>
-						1
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-						disabled
-					>
-						2
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-						disabled
-					>
-						3
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-						disabled
-					>
-						4
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-						disabled
-					>
-						5
-					</button>
-				</div>
-				<div>
-					<button
-						type="button"
-						class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
-						disabled
-					>
-						6
-					</button>
-				</div>
-			</div>
-			<!-- Days -->
+			
 		</div>
 		<!-- End Calendar -->
