+++ swdapp/src/lib/components/date-picker/dummyDatePickerDays.svelte
@@ -1,0 +1,376 @@
+<!-- Days -->
+<div class="flex">
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+            disabled
+        >
+            26
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+            disabled
+        >
+            27
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+            disabled
+        >
+            28
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+            disabled
+        >
+            29
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+            disabled
+        >
+            30
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            1
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            2
+        </button>
+    </div>
+</div>
+<!-- Days -->
+
+<!-- Days -->
+<div class="flex">
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            3
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            4
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            5
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            6
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            7
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            8
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            9
+        </button>
+    </div>
+</div>
+<!-- Days -->
+
+<!-- Days -->
+<div class="flex">
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            10
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            11
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            12
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            13
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            14
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            15
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            16
+        </button>
+    </div>
+</div>
+<!-- Days -->
+
+<!-- Days -->
+<div class="flex">
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            17
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            18
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            19
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent bg-blue-600 text-sm font-medium text-white hover:border-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            20
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            21
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            22
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            23
+        </button>
+    </div>
+</div>
+<!-- Days -->
+
+<!-- Days -->
+<div class="flex">
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            24
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            25
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            26
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            27
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            28
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            29
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            30
+        </button>
+    </div>
+</div>
+<!-- Days -->
+
+<!-- Days -->
+<div class="flex">
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+        >
+            31
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+            disabled
+        >
+            1
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+            disabled
+        >
+            2
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+            disabled
+        >
+            3
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+            disabled
+        >
+            4
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+            disabled
+        >
+            5
+        </button>
+    </div>
+    <div>
+        <button
+            type="button"
+            class="m-px flex size-10 items-center justify-center rounded-full border border-transparent text-sm text-gray-800 hover:border-blue-600 hover:text-blue-600 disabled:pointer-events-none disabled:text-gray-300"
+            disabled
+        >
+            6
+        </button>
+    </div>
+</div>
+<!-- Days -->
